Found a small error in IPFilterSolver. The initial calculation of the neighborhood parameter M was not using alpha_M_initial.
Another thing that needs to be fixed. The IPFilterSolver::Solve method
which has a IPFilterState as argument needs to be corrected. This is
because the provided objective and constraint states f and h in the
IPFilterState instance might not be consistent with the (x,y,z) point if
the optimization problem has changed. Therefore, the correct thing to do
is to provide only the triplet (x,y,z), and not a state.