`BacktrackSearch` no longer uses aggressive Newton stepping mode.
Due to several non robust performance in Reaktoro because of the use of the aggressive Newton stepping approach in Optima, this has been replaced by a conservative Newton step in which mass conservation of elemements are always maintained.

In the aggressive Newton step previously implemented, mass conservation was not ensured in each Newton step. Mass conservation would be violated whenever a chemical species would become attached to the lower bound and the rest of the amounts varying as dictated by the computed Newton step. This caused most equilibrium problems involving H2O-Gypsum-Anhydrite to fail, with H2O becoming completely depleted and Gypsum or Anhydrite becoming many orders of magnitute high (and thus strongly violating mass conservation).

Because of its tendency to fail and also produce extremely high residual errors during the iterative procedure, the aggressive Newton approach will be replaced by the mass-conservative-at-each-Newton-step approach.